Moved spline series into a subfolder.
