+++ js/Series/AreaSplineSeries.js
@@ -9,12 +9,10 @@
  * */
 'use strict';
 import BaseSeries from '../Core/Series/Series.js';
+var areaProto = BaseSeries.seriesTypes.area.prototype;
 import LegendSymbolMixin from '../Mixins/LegendSymbol.js';
 import O from '../Core/Options.js';
 var defaultOptions = O.defaultOptions;
-import './Area/AreaSeries.js';
-import './SplineSeries.js';
-var areaProto = BaseSeries.seriesTypes.area.prototype;
 /* *
  *
  *  Class
