+++ js/Series/Spline/SplineSeries.js
@@ -21,12 +21,10 @@ var __extends = (this && this.__extends) || (function () {
         d.prototype = b === null ? Object.create(b) : (__.prototype = b.prototype, new __());
     };
 })();
-import BaseSeries from '../Core/Series/Series.js';
+import BaseSeries from '../../Core/Series/Series.js';
 var LineSeries = BaseSeries.seriesTypes.line;
-import U from '../Core/Utilities.js';
+import U from '../../Core/Utilities.js';
 var merge = U.merge, pick = U.pick;
-import '../Core/Options.js';
-import '../Series/Line/LineSeries.js';
 /**
  * Spline series type.
  *
