+++ ts/Series/AreaSplineSeries.ts
@@ -19,15 +19,18 @@
 import type AreaPoint from './Area/AreaPoint';
 import type AreaSeries from './Area/AreaSeries';
 import type { SeriesStatesOptions } from '../Core/Series/SeriesOptions';
-import type SplineSeries from './SplineSeries';
+import type SplineSeries from './Spline/SplineSeries';
 import BaseSeries from '../Core/Series/Series.js';
+const {
+    seriesTypes: {
+        area: {
+            prototype: areaProto
+        }
+    }
+} = BaseSeries;
 import LegendSymbolMixin from '../Mixins/LegendSymbol.js';
 import O from '../Core/Options.js';
 const { defaultOptions } = O;
-
-import './Area/AreaSeries.js';
-import './SplineSeries.js';
-var areaProto = BaseSeries.seriesTypes.area.prototype;
 
 /* *
  *
