+++ ts/Series/Line/LineSeries.ts
@@ -40,7 +40,7 @@ import type {
     SeriesTypeOptions,
     SeriesTypePlotOptions
 } from '../../Core/Series/SeriesType';
-import type SplineSeries from '../SplineSeries';
+import type SplineSeries from '../Spline/SplineSeries';
 import type { StatesOptionsKey } from '../../Core/Series/StatesOptions';
 import type SVGAttributes from '../../Core/Renderer/SVG/SVGAttributes';
 import type SVGPath from '../../Core/Renderer/SVG/SVGPath';
